+++ src/concierge_scheduler/__init__.py
@@ -2,7 +2,7 @@
 @author: Gareth Brown
 """
 __all__ = [
-    'concierge_scheduler2',
+    'concierge_scheduler.py',
     'concierge_zabbix',
     'concierge_docker'
 ]
