--- src/concierge_scheduler/docker_env.py
@@ -1,4 +1,0 @@
-DOCKER_CERT_PATH = "/tmp/certs"
-DOCKER_HOST = "tcp://us-east-1.docker.joyent.com:2376"
-DOCKER_CLIENT_TIMEOUT = 800
-COMPOSE_HTTP_TIMEOUT = 800
