--- src/concierge_scheduler/docker_env.sh
@@ -1,4 +1,0 @@
-export DOCKER_CERT_PATH=./
-export DOCKER_HOST=tcp://dockerapi-private-lab1.mesoform.com:2376
-export DOCKER_CLIENT_TIMEOUT=300
-export COMPOSE_HTTP_TIMEOUT=300
